chore: clippy and refactoring
ran clippy
moved `trace_id` and `model` to distinct files
introduced `ANNOTATION_PREFIX` constant
